refactor(dsh-notifier): 按职责拆分宿主入口为 config/message/history/quiet-hours/server 五模块（#18 方案A）
- src/index.ts 1326 行 -> 纯装配层（apply + 事件订阅 + 生命周期）
- config.ts：配置契约/DEFAULT_CONFIG/normalizeConfig/路径函数
- quiet-hours.ts：parseHHMM/isInQuietHours（零依赖叶子，防环形引用）
- message.ts：NOTIFY_KINDS 文案单表/脱敏有序表/格式化/agent 事件读取
- history.ts：jsonl 存储工厂（滚动上限/按天清理/tmp+rename 原子写/写队列串行化）
- server.ts：SSE 枢纽/系统通知通道/buildRoutes 路由组装
- 导出面不变：运行时与 d.ts 符号集与拆分前逐一比对一致；
  appendHistory/systemNotify/SSE 等闭包逻辑参数化时改传实时 getter
  （maxAgeDays/notifySound），保持 PUT /config 后立即生效的语义
