+++ packages/dsh-notifier/src/index.ts
@@ -1,6 +1,13 @@
 /**
  * dsh-notifier — 审批/完成/错误事件通知（宿主端）。
  *
+ * 本文件只做装配：apply + 事件订阅 + 生命周期清理。职责划分：
+ * - config.ts     配置契约/默认值/normalizeConfig/路径
+ * - quiet-hours.ts 免打扰判定（parseHHMM / isInQuietHours）
+ * - message.ts    通知文案单表/脱敏/格式化/agent 事件读取（纯函数）
+ * - history.ts    通知历史 jsonl 存储（滚动/按天清理/原子写）
+ * - server.ts     SSE 枢纽 + 系统通知通道 + HTTP 路由
+ *
  * 事件源（均经源码核验）：
  * - 审批请求：approval/request waterfall（ApprovalService.request 派发；
  *   真实触发方为沙箱提权 / 动态插件激活等。tools/pre-execute 的 gate ask
@@ -23,27 +30,24 @@
  * 配置：~/.dsh/dsh-notifier.json（PUT /config 可改，落盘）
  * 安全：通知文本只含工具名/会话标识/申请理由等元信息，不含工具参数（防敏感信息外泄）。
  */
-import { spawn, execFile } from "node:child_process";
-import type { ChildProcess } from "node:child_process";
-import { readFile, writeFile, mkdir, rename } from "node:fs/promises";
-import { existsSync } from "node:fs";
-import { homedir } from "node:os";
-import { dirname, join } from "node:path";
-import { fileURLToPath } from "node:url";
-import type { ServerResponse } from "node:http";
 // 官方类型层（issue #16，锁 0.1.1-rc.2；仅 import type，编译期擦除，
 // 禁止运行时值导入——contract-check 有门禁）。session/title 事件键由
 // dsh-session-title 经 declare module 注入 SessionEventMap，必须引入其类型面。
+import { readFile, writeFile, mkdir, rename } from "node:fs/promises";
+import { existsSync } from "node:fs";
+import { dirname } from "node:path";
 import type { Context } from "@deepseek-ai/cordis";
 import type { Agent } from "@deepseek-ai/dsh-agent";
-import type { WebRoute } from "@deepseek-ai/dsh-host-webserver";
-import type { TurnEndReason } from "@deepseek-ai/dsh-session/types";
 import type {} from "@deepseek-ai/dsh-session-title";
 import type {} from "@deepseek-ai/dsh-user-approval";
-
-/** turn/end reason.kind 的官方联合（TurnEndReasonMap；插件可经 declare module 扩展，
- * 运行时出现未知 kind 由调用方保守静默——见 agent/status 完成判定白名单）。 */
-type TurnEndKind = TurnEndReason extends { kind: infer K } ? K : never;
+import { errorMessage } from "../../../shared/host-utils.js";
+import { CONFIG_KEYS, DEFAULT_CONFIG, configFile, historyFile, normalizeConfig, toastScriptPath } from "./config.js";
+import type { NotifierApplyConfig, NotifyConfig } from "./config.js";
+import { isInQuietHours } from "./quiet-hours.js";
+import { HISTORY_LIMIT, createHistoryStore } from "./history.js";
+import { NOTIFY_KINDS, sanitizeErrorText, sessionTitleOf, isSubagentOf, lastTurnEndOf } from "./message.js";
+import type { NotifyDetail } from "./message.js";
+import { ROUTES, buildRoutes, createSseHub, createSystemNotifier } from "./server.js";
 
 /** 稳定的 cordis 插件名。 */
 export const name = "notifier";
@@ -51,443 +55,29 @@ export const name = "notifier";
 /** 需要的服务：webServer（路由）。 */
 export const inject = ["webServer"];
 
-/** 免打扰时段配置。 */
-export interface QuietHoursConfig {
-  enabled: boolean;
-  start: string;
-  end: string;
-  /** 免打扰期间仍放行的通知 kind（紧急例外，如审批/提问——卡着的任务需要叫醒）。 */
-  allowKinds?: string[];
-}
+// ---------------------------------------------------------------- 导出面
+// 公共符号定义在各职责模块，此处统一 re-export——包导出面与拆分前完全一致。
 
-/** 免打扰紧急例外可选 kind（面板展示与此白名单一致）。 */
-export const QUIET_ALLOW_KINDS = ["ask", "question", "error"] as const;
-
-/** 通知配置（内存单一事实源，与落盘 JSON 同构）。 */
-export interface NotifyConfig {
-  notifyAsk: boolean;
-  notifyQuestion: boolean;
-  notifyTaskDone: boolean;
-  /** 子代理完成通知（独立于主任务完成，默认关）。 */
-  notifySubagentDone: boolean;
-  notifyTaskError: boolean;
-  notifyTurnEnd: boolean;
-  systemNotify: boolean;
-  browserNotify: boolean;
-  notifyWhenVisible: boolean;
-  /** 系统通知是否带提示音（false = silent，静默弹出）。 */
-  notifySound: boolean;
-  quietHours: QuietHoursConfig;
-  errorMergeWindowMs: number;
-  /** 审批等待超时二次提醒（分钟；0 = 关闭）。 */
-  askRemindMin: number;
-  /** 完成风暴聚合窗口（毫秒；0 = 关闭聚合，每条完成即时通知）。 */
-  doneMergeWindowMs: number;
-  /** 通知历史按天自动清理（0 = 不按时间清理，仅行数上限滚动）。 */
-  historyMaxAgeDays: number;
-}
-
-/** 布尔配置键联合（normalizeConfig 白名单与客户端渲染依赖它）。 */
-type BooleanKeys = { [K in keyof NotifyConfig]: NotifyConfig[K] extends boolean ? K : never }[keyof NotifyConfig];
-
-/** apply 接收的配置（enabled / 路径覆盖）。 */
-export interface NotifierApplyConfig {
-  enabled?: boolean;
-  configFile?: string;
-  toastScript?: string;
-  historyFile?: string;
-}
-
-/** 通知详情（不含工具参数等敏感信息）。 */
-export interface NotifyDetail {
-  tool?: string;
-  taskTitle?: string;
-  reason?: string;
-  question?: string;
-  durationMs?: number;
-  turn?: number;
-  step?: number;
-  message?: string;
-  mergedCount?: number;
-  /** 审批超时提醒：已等待分钟数（NOTIFY_KINDS.ask 渲染）。 */
-  remindMinutes?: number;
-  /** 错误合并窗口内被吞掉的错误摘要（最近 2 条）。 */
-  mergedErrors?: string[];
-  ts?: number;
-}
-
-/** Agent 事件 payload 的类型来源：官方 @deepseek-ai/dsh-agent（原手写 AgentLite
- * 已删，字段语义见官方 SessionHeader.origin / Session.events 判别联合）。 */
-
-/** 与客户端共享的路由常量（smoke 断言两端一致）。 */
-export const ROUTES = {
-  config: "/api/dsh-notifier/config",
-  events: "/api/dsh-notifier/events",
-  health: "/api/dsh-notifier/health",
-  test: "/api/dsh-notifier/test",
-  history: "/api/dsh-notifier/history",
-};
-
-/** 通知历史滚动上限（行数；超出后从尾部截断重写）。 */
-export const HISTORY_LIMIT = 200;
-
-/** 通知历史文件路径（jsonl 追加；与配置同目录）。 */
-export function historyFile() {
-  return join(homedir(), ".dsh", "dsh-notifier-history.jsonl");
-}
-
-/** 默认配置。 */
-export const DEFAULT_CONFIG: NotifyConfig = {
-  notifyAsk: true,
-  /** 用户提问（ask_user_question / GUI 提问弹窗）时通知。 */
-  notifyQuestion: true,
-  notifyTaskDone: true,
-  /** 子代理完成通知（独立开关，默认关：派生任务完成不打扰主流程）。 */
-  notifySubagentDone: false,
-  notifyTaskError: true,
-  notifyTurnEnd: false,
-  systemNotify: true,
-  browserNotify: true,
-  /** 页面可见（聚焦）时也弹浏览器通知；默认 false = 仅页面隐藏时弹（避免打扰）。 */
-  notifyWhenVisible: false,
-  /** 系统通知默认带提示音。 */
-  notifySound: true,
-  quietHours: { enabled: false, start: "22:00", end: "08:00" },
-  /** 同类错误合并窗口（毫秒）：窗口内后续错误不再单独通知，累计到下次一并提示。 */
-  errorMergeWindowMs: 60000,
-  /** 审批等待超时二次提醒（分钟，0=关闭；审批被卡是最高成本事件，值得重提醒）。 */
-  askRemindMin: 5,
-  /** 完成风暴聚合窗口（毫秒；0=关闭——并行子代理收尾防刷屏，代价是窗口内
-   *  后续完成会延迟到窗口到点以聚合条补发）。 */
-  doneMergeWindowMs: 3000,
-  /** 通知历史按天自动清理（0=不按时间清理，仅靠行数上限滚动）。 */
-  historyMaxAgeDays: 0,
-};
-
-/** 布尔配置键（单一事实源：normalizeConfig 白名单与客户端渲染依赖它）。 */
-export const CONFIG_KEYS: readonly BooleanKeys[] = ["notifyAsk", "notifyQuestion", "notifyTaskDone", "notifySubagentDone", "notifyTaskError", "notifyTurnEnd", "systemNotify", "browserNotify", "notifyWhenVisible", "notifySound"];
-
-/** 配置存储路径。 */
-export function configFile() {
-  return join(homedir(), ".dsh", "dsh-notifier.json");
-}
-
-/** toast 脚本路径（本插件 lib 下）。 */
-export function toastScriptPath() {
-  return join(dirname(fileURLToPath(import.meta.url)), "toast.ps1");
-}
-
-/** 合并配置（未知键丢弃，默认值兜底；深拷贝防默认值被污染）。 */
-export function normalizeConfig(input: unknown): NotifyConfig {
-  const base: NotifyConfig = { ...DEFAULT_CONFIG, quietHours: { ...DEFAULT_CONFIG.quietHours } };
-  if (typeof input !== "object" || input === null) return base;
-  const src = input as Record<string, unknown>;
-  for (const key of CONFIG_KEYS) {
-    if (typeof src[key] === "boolean") base[key] = src[key];
-  }
-  if (Number.isFinite(src.errorMergeWindowMs) && (src.errorMergeWindowMs as number) >= 0 && (src.errorMergeWindowMs as number) <= 3600000) {
-    base.errorMergeWindowMs = Math.round(src.errorMergeWindowMs as number);
-  }
-  if (Number.isFinite(src.askRemindMin) && (src.askRemindMin as number) >= 0 && (src.askRemindMin as number) <= 600) {
-    base.askRemindMin = Math.round(src.askRemindMin as number);
-  }
-  if (Number.isFinite(src.doneMergeWindowMs) && (src.doneMergeWindowMs as number) >= 0 && (src.doneMergeWindowMs as number) <= 60000) {
-    base.doneMergeWindowMs = Math.round(src.doneMergeWindowMs as number);
-  }
-  if (Number.isFinite(src.historyMaxAgeDays) && (src.historyMaxAgeDays as number) >= 0 && (src.historyMaxAgeDays as number) <= 3650) {
-    base.historyMaxAgeDays = Math.round(src.historyMaxAgeDays as number);
-  }
-  if (typeof src.quietHours === "object" && src.quietHours !== null) {
-    const qh = src.quietHours as Record<string, unknown>;
-    if (typeof qh.enabled === "boolean") base.quietHours.enabled = qh.enabled;
-    // 时/分范围校验（0-23/0-59）：复用 parseHHMM，非法值（如 25:00）丢弃回默认，
-    // 避免「配置保存成功但免打扰永不生效」的静默失败
-    if (typeof qh.start === "string" && /^\d{2}:\d{2}$/u.test(qh.start) && parseHHMM(qh.start) >= 0) base.quietHours.start = qh.start;
-    if (typeof qh.end === "string" && /^\d{2}:\d{2}$/u.test(qh.end) && parseHHMM(qh.end) >= 0) base.quietHours.end = qh.end;
-    // 紧急例外 kind 白名单过滤
-    if (Array.isArray(qh.allowKinds)) {
-      base.quietHours.allowKinds = qh.allowKinds.filter((k): k is string => typeof k === "string" && (QUIET_ALLOW_KINDS as readonly string[]).includes(k));
-    }
-  }
-  // 未知键透传：白名单之外的键原样保留（此插件在旧版本运行或手改配置时会
-  // 出现未来版本/第三方键），避免「降级丢键」——只归一化你认识的键。
-  const out = base as NotifyConfig & Record<string, unknown>;
-  for (const key of Object.keys(src)) {
-    if ((CONFIG_KEYS as readonly string[]).includes(key)) continue;
-    // 已归一化过的键不再透传（否则非法值会以原样覆盖归一化结果）
-    if (key === "quietHours" || key === "errorMergeWindowMs" || key === "askRemindMin" || key === "doneMergeWindowMs" || key === "historyMaxAgeDays") continue;
-    out[key] = src[key];
-  }
-  return out;
-}
-
-/** 毫秒 → 人类可读耗时（如 "45 秒" / "2 分 15 秒" / "1 小时 2 分 5 秒"）。 */
-export function formatDuration(ms: number): string {
-  const totalSeconds = Math.max(0, Math.round(Number(ms) / 1000));
-  const hours = Math.floor(totalSeconds / 3600);
-  const minutes = Math.floor((totalSeconds % 3600) / 60);
-  const seconds = totalSeconds % 60;
-  const parts: string[] = [];
-  if (hours > 0) parts.push(`${hours} 小时`);
-  if (minutes > 0) parts.push(`${minutes} 分`);
-  if (seconds > 0 || parts.length === 0) parts.push(`${seconds} 秒`);
-  return parts.join(" ");
-}
-
-/**
- * 脱敏规则有序表：顺序即数据（具体在前、泛化在后），消灭隐式语句序。
- * 顺序硬约束（对抗评审实测确立）：
- * - DSN 必须先于 JWT/通用长串/密钥赋值：否则 DSN 密码先被劣化为 <token>，
- *   DSN 字符类排除 <> 随即失配，用户名残留明文；
- * - 邮箱必须殿后且带 (?<!<) 界定：避免把 DSN 掩码占位符 <redacted>@host
- *   中的 "redacted@host" 再误判为邮箱产生 <<email>>。
- * 已证伪不收录（勿加回）：IPv4（UA 版本号 Chrome/120.0.0.0 形态不可区分，
- * 误伤高频）、手机号（订单号误伤、错误文本出现概率趋零）、信用卡
- * （13 位毫秒时间戳 100% 命中，Luhn 校验也救不了）。
- */
-const SANITIZE_RULES: ReadonlyArray<readonly [RegExp, string]> = [
-  // 用户路径（/home/xx、/Users/xx、/root/xx、/etc/xx、C:\Users\xx）→ <path>
-  // （排除冒号/分号等后缀分隔符，避免吞掉紧随其后的文本：`/home/x: EACCES` 应保留冒号）
-  [/(?:\/home\/[^\s"'`<>:;]+|\/Users\/[^\s"'`<>:;]+|\/root\/[^\s"'`<>:;]+|\/etc\/[^\s"'`<>:;]+|C:\\Users\\[^\s"'`<>:;]+)/giu, "<path>"],
-  // PEM 私钥完整块（RSA/EC/OPENSSH/ENCRYPTED…）→ <private-key>
-  // 窗口上限 {0,4096}：真实私钥块 ≤ 数 KB；无窗口时每个 BEGIN 起点都会惰性扫到
-  // 文本尾找 END，k 个 BEGIN 即 O(k·n)，1MB 高密度输入可阻塞宿主事件循环数秒
-  // （评审实测）。超窗视为无 END，交由下一条孤立 BEGIN 兜底规则接住。
-  [/-----BEGIN [A-Z ]*PRIVATE KEY-----[\s\S]{0,4096}?-----END [A-Z ]*PRIVATE KEY-----/g, "<private-key>"],
-  // PEM 孤立 BEGIN 兜底：错误消息常只贴出头几行就被截断，无 END 行；
-  // 吃到文本尾即可（尾部反正受 maxLen 截断）
-  [/-----BEGIN [A-Z ]*PRIVATE KEY-----[\s\S]*/g, "<private-key>"],
-  // 数据库/消息队列连接串凭据 → 只留 scheme://<redacted>@host
-  // 用户名允许为空（redis://:pass@host 是最常见带密码形态）；i flag 兼容大写 scheme；
-  // amqps?：AMQPS 同端口语义，漏收会产生「密码被邮箱规则侥幸掩掉、用户名明文残留」
-  // 的半脱敏误导形态；jdbc 不收录：JDBC 凭据在 query 参数（?password=），密钥赋值规则已覆盖。
-  // 已知局限（评审记录）：密码含 <>/引号等 URL 应编码字符时整条失配、明文残留——
-  // 错误消息中的未编码形态不罕见，但放宽字符类会引入占位符/引号边界歧义，暂记录不改。
-  [/\b(postgres(?:ql)?|mysql|mariadb|mongodb(?:\+srv)?|rediss?|amqps?|mssql):\/\/([^\s:@\/"'`<>]*):([^\s@\/"'`<>]+)@/giu, "$1://<redacted>@"],
-  // GitHub PAT classic（ghp/gho/ghu/ghs/ghr + 36 位字母数字；业界共识写死长度）
-  [/\b(?:ghp|gho|ghu|ghs|ghr)_[A-Za-z0-9]{36}\b/gu, "<token>"],
-  // GitHub fine-grained PAT（github_pat_ 强前缀锚定，长度取区间吸收未来调整）
-  [/\bgithub_pat_[A-Za-z0-9]{20,30}_[A-Za-z0-9]{50,80}\b/gu, "<token>"],
-  // JWT（eyJ JWT：header.payload.signature 三段，base64url 字符集含 - _）→ <token>
-  // 精确匹配三段（每段 ≥1 字符，含 - _），避免误伤普通长串（对抗评审建议：勿盲目放宽
-  // 通用 base64 正则到 base64url，那会误伤 URL/长串）。
-  [/\beyJ[A-Za-z0-9_-]+(\.[A-Za-z0-9_-]+){2}\b/gu, "<token>"],
-  // AWS 访问密钥前缀（AKIA 20 字符，不满足通用 ≥24hex/≥32base64 阈值）→ <token>
-  [/\bAKIA[A-Z0-9]{16}\b/gu, "<token>"],
-  // 长令牌/密钥串（≥24 hex 或 ≥32 base64）→ <token>
-  [/\b(?:[0-9a-fA-F]{24,}|[A-Za-z0-9+/]{32,}={0,2})\b/gu, "<token>"],
-  // 密钥字段赋值（password=/token=/api_key=…）→ 只留键名+掩码
-  // 分隔符 [=:] 必须显式：可选版会把 "token expired"/"password policy" 等自然语言
-  // 高频误伤成 token=<redacted>（审批理由/提问文本接入后误伤面被放大，评审实测）
-  [/\b(password|passwd|token|api[_-]?key|secret|authorization)\b\s*[=:]\s*["']?[^\s"'`,;<>]{3,}/giu, "$1=<redacted>"],
-  // 邮箱（严格版：域名首标签须字母开头，排除 image@2x.png / pkg@1.2.3 误伤）→ <email>
-  // 负向后行断言排除词字符与 <：防止从 <redacted>@host 占位符中间起配
-  [/(?<![A-Za-z0-9._%+<>-])[A-Za-z0-9._%+-]+@[A-Za-z][A-Za-z0-9.-]*\.[A-Za-z]{2,}/gu, "<email>"],
-];
-
-/**
- * 错误文本脱敏：按 SANITIZE_RULES 有序表掩蔽常见敏感特征（用户路径、私钥、
- * 连接串凭据、各类令牌、密钥赋值、邮箱），再截断。
- * 用于 agent/error 与审批理由/提问文本进入通知与历史前的处理，把「文本可能
- * 内嵌命令回显/路径/凭据片段」的外泄面收敛到可读的摘要。
- * @returns 脱敏并截断（默认 300 字符）后的错误文本。
- */
-export function sanitizeErrorText(text: unknown, maxLen = 300): string {
-  let s = String(text);
-  for (const [pattern, replacement] of SANITIZE_RULES) {
-    s = s.replace(pattern, replacement);
-  }
-  return s.slice(0, maxLen);
-}
-
-/**
- * 构造系统通知命令参数（纯函数，smoke 可直接断言参数形态）。
- * - win32：spawn powershell -File toast.ps1 -Title=<t> -Message=<m> [-Silent]
- *   单 token 传参：封掉「消息以 - 开头被解析成参数名」与含空格的歧义，
- *   依旧零 shell 拼接面
- * - darwin：osascript display notification（转义 \ 与 "，换行替换为空格防
- *   脚本语法；silent=false 时带系统提示音 "Glass"）
- * - 其余：notify-send（notifySendAvailable === false 时返回 null=通道不可用）
- * @returns spawn 参数数组（首元素为可执行文件），或 null（不可用）。
- */
-export function buildSystemCommand(
-  platform: string,
-  title: string,
-  message: string,
-  options: { silent: boolean; notifySendAvailable?: boolean; toastScript: string }
-): string[] | null {
-  if (platform === "win32") {
-    const args = ["-NoProfile", "-NonInteractive", "-ExecutionPolicy", "Bypass", "-File", options.toastScript, `-Title=${title}`, `-Message=${message}`];
-    if (options.silent) args.push("-Silent");
-    return ["powershell", ...args];
-  }
-  if (platform === "darwin") {
-    const esc = (s: string) => s.replace(/\\/gu, "\\\\").replace(/"/gu, '\\"').replace(/\n/gu, " ");
-    const script = `display notification "${esc(message)}" with title "${esc(title)}"` + (options.silent ? "" : ' sound name "Glass"');
-    return ["osascript", "-e", script];
-  }
-  if (options.notifySendAvailable === false) return null;
-  return ["notify-send", title, message];
-}
-
-/** 常见工具名 → 中文可读名（业界通知惯例：用用户看得懂的动作描述，而非内部标识）。 */
-const TOOL_LABELS: Record<string, string> = {
-  pwsh: "PowerShell 命令",
-  bash: "终端命令",
-  "web_search": "联网搜索",
-  read: "读取文件",
-  write: "写入文件",
-  edit: "编辑文件",
-  grep: "搜索文件内容",
-  glob: "查找文件",
-  subagent: "子代理任务",
-  subagent_fork: "子代理任务",
-  ssh_exec: "SSH 远程执行",
-  ssh_upload: "SSH 上传文件",
-  ssh_download: "SSH 下载文件",
-  ssh_tunnel: "SSH 端口转发",
-  ssh_cluster: "SSH 集群执行",
-  ask_user_question: "向你提问",
-  todo_write: "更新任务清单",
-  job_list: "查看后台任务",
-  job_output: "查看任务输出",
-  job_kill: "停止后台任务",
-  workflow: "工作流编排",
-  ralph: "Ralph 循环执行",
-  memory_add: "写入项目记忆",
-  memory_search: "检索项目记忆",
-  skill: "加载技能",
-  create_goal: "创建目标",
-  get_goal: "查看目标",
-  update_goal: "更新目标",
-};
-
-/**
- * 工具名美化（业界惯例：可读的动作描述优先）：
- * - 常见工具映射表 → 中文名（如 pwsh → "PowerShell 命令"）
- * - `mcp__server__raw` → `MCP 服务器 "server" 的工具 "raw"`
- * - 其余原样
- */
-export function prettyToolName(name: unknown): string {
-  const text = String(name ?? "?");
-  if (Object.prototype.hasOwnProperty.call(TOOL_LABELS, text)) return TOOL_LABELS[text];
-  const parts = text.split("__");
-  if (parts.length >= 3 && parts[0] === "mcp") {
-    return `MCP 服务器 "${parts[1]}" 的工具 "${parts.slice(2).join("__")}"`;
-  }
-  return text;
-}
-
-/**
- * 提取会话标题（用户可读的任务名，替代内部 session id）。
- * 数据源：agent.session.events 中最后一个 session/title 事件（dsh-session-title
- * 官方插件维护，与 GUI 会话列表同源）。无标题（新会话/未生成）返回 undefined。
- * @param agent Agent 对象（事件 payload.agent）。
- * @returns 截断 40 字符的标题。
- */
-export function sessionTitleOf(agent: Agent | undefined): string | undefined {
-  try {
-    // 运行时防御保留（payload 跨宿主边界，不受信）：显式收窄而非 Array.isArray
-    // （后者会把 readonly SessionEvent[] 压成 any[]，令 typecheck 对 events 访问失明）。
-    const events = agent?.session?.events as ReadonlyArray<{ type: unknown; data?: { title?: unknown } }> | undefined;
-    if (events === undefined) return undefined;
-    for (let i = events.length - 1; i >= 0; i -= 1) {
-      const ev = events[i] as { type: unknown; data?: { title?: unknown } } | undefined;
-      if (ev?.type === "session/title" && typeof ev.data?.title === "string") {
-        const title = ev.data.title.trim();
-        return title.length > 0 ? title.slice(0, 40) : undefined;
-      }
-    }
-  } catch {
-    // 事件日志读取失败不影响通知主流程
-  }
-  return undefined;
-}
-
-/**
- * 取 agent 会话日志中最新一条 turn/end（倒序扫描；session.events 是混合
- * 日志，turn/end 后可能尾随 session/title、inbox、user/message 等，不能
- * 取 events 末尾——照 sessionTitleOf 同款倒序）。
- * 用于中断抑制：running→idle 时若本轮未闭合新的 turn/end（如 abort 早于
- * turn/start 落盘，dsh-agent-loop turn() 首行 throwIfAborted 在 try 外），
- * 读到的 turn/end 是上一次运行的，不可作为本轮结束依据。
- * @param agent Agent 对象（事件 payload.agent）。
- * @returns {turn, kind}（kind 为 turn/end reason.kind，如 completed /
- *   aborted / error / max-tokens）；无 turn/end 或读取失败返回 undefined。
- */
-export function lastTurnEndOf(agent: Agent | undefined): { turn: number; kind: TurnEndKind } | undefined {
-  try {
-    // 运行时防御同 sessionTitleOf：显式收窄，不用 Array.isArray（any 化陷阱）
-    const events = agent?.session?.events as ReadonlyArray<{ type: unknown; data?: { turn?: unknown; reason?: unknown } }> | undefined;
-    if (events === undefined) return undefined;
-    for (let i = events.length - 1; i >= 0; i -= 1) {
-      const ev = events[i] as { type: unknown; data?: { turn?: unknown; reason?: unknown } } | undefined;
-      if (ev?.type !== "turn/end") continue;
-      const reason = ev.data?.reason;
-      if (reason === undefined || reason === null || typeof reason !== "object") continue;
-      const turn = typeof ev.data?.turn === "number" ? ev.data.turn : NaN;
-      const kind = String((reason as { kind?: unknown }).kind ?? "") as TurnEndKind;
-      return { turn, kind };
-    }
-  } catch {
-    // 事件日志读取失败不影响通知主流程
-  }
-  return undefined;
-}
-
-/**
- * agent 是否为子代理：以 session.header.origin === 'subagent' 为唯一判据。
- *
- * 选此信号而非 delegationDepth / parentSession 的原因（均经 DSH 源码核验）：
- * - origin 由 SessionHeader 校验强制唯一合法值为 'subagent'
- *   （packages/core/session/src/index.ts:125：origin 非 undefined 就必须是
- *   'subagent'），主会话/fork 派生会话恒为 undefined → 零误判面。
- * - DSH 自身在无父 id 场景（packages/subagent/subagent/src/list-children.ts:324、
- *   :235）也只认 origin === 'subagent' 识别子代理，口径一致。
- * - parentSession 不能作判据：Session.fork()（session/src/index.ts:1091）与
- *   apiproxy 的 lineage 场景会写 parentSession 但不带 origin → 若用 parentSession
- *   会把 fork/派生主流程会话误判为子代理，其「任务完成」被默认
- *   notifySubagentDone=false 静默（当前 bug 的镜像回归）。
- * - delegationDepth 在某些路径（resume 信任持久化 header、运行时 deepen）下并非
- *   可靠的「是否子代理」标记，仅作兜底。
- * 综上：单用 origin 最权威、最稳；notifier 是单 agent 独立判定，不要求父 id 匹配。
- */
-export function isSubagentOf(agent: Agent | undefined): boolean {
-  return agent?.session?.header?.origin === "subagent";
-}
-
-/** "HH:MM" → 分钟数（校验 0-23 时 / 0-59 分）。 */
-export function parseHHMM(text: string): number {
-  const m = /^(\d{2}):(\d{2})$/u.exec(text);
-  if (m === null) return NaN;
-  const hours = Number(m[1]);
-  const minutes = Number(m[2]);
-  if (hours > 23 || minutes > 59) return NaN;
-  return hours * 60 + minutes;
-}
-
-/** 是否处于免打扰时段（支持跨午夜：start > end）。 */
-export function isInQuietHours(now: Date, quietHours: QuietHoursConfig | undefined): boolean {
-  if (quietHours?.enabled !== true) return false;
-  const minutes = now.getHours() * 60 + now.getMinutes();
-  const start = parseHHMM(quietHours.start);
-  const end = parseHHMM(quietHours.end);
-  if (Number.isNaN(start) || Number.isNaN(end)) return false;
-  if (start === end) return false;
-  if (start < end) return minutes >= start && minutes < end;
-  return minutes >= start || minutes < end;
-}
+export { QUIET_ALLOW_KINDS, isInQuietHours, parseHHMM } from "./quiet-hours.js";
+export type { QuietHoursConfig } from "./quiet-hours.js";
+export { CONFIG_KEYS, DEFAULT_CONFIG, configFile, historyFile, normalizeConfig, toastScriptPath } from "./config.js";
+export type { NotifierApplyConfig, NotifyConfig } from "./config.js";
+export { HISTORY_LIMIT } from "./history.js";
+export {
+  buildSystemCommand,
+  formatDuration,
+  isSubagentOf,
+  lastTurnEndOf,
+  prettyToolName,
+  sanitizeErrorText,
+  sessionTitleOf,
+} from "./message.js";
+export type { NotifyDetail } from "./message.js";
+export { ROUTES } from "./server.js";
 
 // 辅助函数统一来自仓库共享层（loopback 围栏 / writeJson / readBody / errorMessage）。
-import { isLoopbackRequest } from "../../../shared/loopback.js";
-import { writeJson, readBody, errorMessage } from "../../../shared/host-utils.js";
 export { isLoopbackRequest } from "../../../shared/loopback.js";
 export { writeJson, readBody, errorMessage } from "../../../shared/host-utils.js";
-
-/** SSE 帧。 */
-function sseData(payload: unknown): string {
-  return `data: ${JSON.stringify(payload)}\n\n`;
-}
 
 /**
  * 挂载 dsh-notifier。
@@ -498,7 +88,7 @@ export async function apply(ctx: Context, config: NotifierApplyConfig = {}): Pro
   if (config.enabled === false) return;
   const storeFile = typeof config.configFile === "string" ? config.configFile : configFile();
   const toastScript = typeof config.toastScript === "string" ? config.toastScript : toastScriptPath();
-  const historyStore = typeof config.historyFile === "string" ? config.historyFile : historyFile();
+  const historyPath = typeof config.historyFile === "string" ? config.historyFile : historyFile();
 
   /** 当前配置（内存单一事实源，PUT 后落盘）。 */
   let current = normalizeConfig(undefined);
@@ -524,239 +114,26 @@ export async function apply(ctx: Context, config: NotifierApplyConfig = {}): Pro
     }
   }
 
-  // ------------------------------------------------------------ 系统通知
+  // ------------------------------------------------------------ 组装各职责模块
 
-  /** 系统通知节流间隔：防连发（生产密集事件/连点测试按钮）造成 spawn 风暴。 */
-  const SYSTEM_NOTIFY_THROTTLE_MS = 1000;
-  let lastSystemNotifyAt = 0;
+  const sse = createSseHub();
+  const system = createSystemNotifier({
+    getSoundEnabled: () => current.notifySound,
+    toastScript,
+    warn: (message) => ctx.logger.warn(message),
+  });
+  const historyStore = createHistoryStore({
+    file: historyPath,
+    maxAgeDays: () => current.historyMaxAgeDays,
+    warn: (message) => ctx.logger.warn(message),
+  });
 
-  /**
-   * 系统原生 toast（fire-and-forget + 30s 超时杀进程；全部失败静默）。
-   * 命令构造走顶层纯函数 buildSystemCommand（smoke 断言参数形态）：
-   * Windows：PowerShell WinRT toast；macOS：osascript display notification；
-   * 其他：notify-send（可用才调用）。
-   */
-  function systemNotify(title: string, message: string) {
-    // 1s 节流：桌面通知连发只保留第一个，避免子进程风暴
-    const now = Date.now();
-    if (now - lastSystemNotifyAt < SYSTEM_NOTIFY_THROTTLE_MS) return;
-    lastSystemNotifyAt = now;
-    const safeTitle = String(title).slice(0, 64);
-    const safeMessage = String(message).slice(0, 256);
-    let child: ChildProcess | undefined;
-    let bin = "";
-    try {
-      const argv = buildSystemCommand(process.platform, safeTitle, safeMessage, {
-        silent: current.notifySound === false,
-        notifySendAvailable,
-        toastScript,
-      });
-      if (argv === null) return; // 通道不可用（如 Linux 探测到 notify-send 缺失）：静默跳过
-      bin = argv[0];
-      // 关键：原生二进制缺失/不可执行（ENOENT 等）必须被下方 error 事件接住，
-      // 绝不能冒泡成 unhandled 'error' 把宿主进程打挂——历史版本在 macOS 上因
-      // 直接 spawn powershell 失败且未挂 error 监听而崩溃（见 issue #1）。
-      child = spawn(bin, argv.slice(1), process.platform === "win32" ? { windowsHide: true, stdio: "ignore" } : { stdio: "ignore" });
-    } catch (error) {
-      ctx.logger.warn(`dsh-notifier: 系统通知启动失败: ${errorMessage(error)}`);
-      return;
-    }
-    if (!child) return; // 极端情况下 spawn 返回空（理论上不会），直接放弃，不碰 child
-    const killer = setTimeout(() => {
-      try {
-        child!.kill();
-      } catch {
-        // 忽略
-      }
-    }, 30000);
-    // 任何退出码都先清杀手定时器；非 0 且非被信号杀死（null）才记日志，
-    // 避免把「我们主动 30s 超时杀掉子进程」也当成异常刷屏。
-    child.on("exit", (code) => {
-      clearTimeout(killer);
-      if (code !== 0 && code !== null) {
-        ctx.logger.warn(`dsh-notifier: 系统通知退出码异常（exit ${code}）`);
-      }
-    });
-    // 通道失败（二进制不在 PATH / 无桌面会话等）：非 Windows 记一次后置为不可用，
-    // 后续通知走 null 分支静默跳过；Windows 不置位（powershell 偶发失败不代表
-    // WinRT toast 通道整体失效）。无论如何只记日志、不抛、不崩。
-    child.on("error", (error) => {
-      clearTimeout(killer);
-      if (process.platform !== "win32") notifySendAvailable = false;
-      ctx.logger.warn(`dsh-notifier: 系统通知不可用（${bin}）: ${errorMessage(error)}`);
-    });
+  /** 通知历史追加（fire-and-forget 落盘，见 history.ts）。 */
+  function appendHistory(entry: { ts: number; kind: string; title: string; message: string; suppressed?: string }) {
+    historyStore.append(entry);
   }
-
-  /** notify-send 可用性探测（仅 Linux 需要；macOS 走 osascript，darwin 分支
-   *  不依赖此探测，故不在 macOS 上无谓尝试缺失的 notify-send）。异步，只探一次。 */
-  let notifySendAvailable: boolean | undefined = undefined;
-  if (process.platform === "linux") {
-    execFile("notify-send", ["--version"], { timeout: 3000 }, (error) => {
-      notifySendAvailable = error === null;
-    });
-  }
-
-  // ------------------------------------------------------------ SSE 推送
-
-  const sseConnections = new Set<ServerResponse>();
-
-  /** SSE 已派发帧的滚动缓冲（断线回补用；上限 RECENT_LIMIT，独立于 /history 的
-   *  200 条截断，避免补拉时尾部事件被截掉）。 */
-  const RECENT_LIMIT = 600;
-  let notifySeq = 0;
-  const recentFrames: Array<Record<string, unknown> & { seq: number }> = [];
-
-  function broadcast(payload: Record<string, unknown>) {
-    const frame: Record<string, unknown> & { seq: number } = Object.assign({}, payload, { seq: ++notifySeq });
-    recentFrames.push(frame);
-    if (recentFrames.length > RECENT_LIMIT) recentFrames.shift();
-    const text = sseData(frame);
-    for (const res of sseConnections) {
-      try {
-        res.write(text);
-      } catch {
-        // 连接已断，等待 close 清理
-      }
-    }
-  }
-
-  /**
-   * SSE 心跳：每 30s 发一条 data 型 ping 帧。必须用 data 而非注释帧——注释帧
-   * 既不触发客户端 onmessage 也不触发 onerror，半开连接（断网/合盖/NAT 静默
-   * 掐断）时客户端零事件，无法自愈；data 帧让客户端的 watchdog 能检测失活。
-   */
-  const HEARTBEAT_MS = 30000;
-  const heartbeatTimer = setInterval(() => {
-    const frame = sseData({ type: "ping" });
-    for (const res of sseConnections) {
-      try {
-        res.write(frame);
-      } catch {
-        // 等待 close 清理
-      }
-    }
-  }, HEARTBEAT_MS);
-  // unref：心跳不阻止进程退出（服务端有主循环；测试/无连接时不会被句柄吊死）
-  heartbeatTimer.unref();
 
   // ------------------------------------------------------------ 通知主入口
-
-  // 通知文案单表（kind → {title, message}），收敛 TITLES 与 switch 双映射。
-  // 设计遵循业界通知惯例：标题即结论、正文结论先行、用用户可读的任务名与
-  // 动作描述（不暴露内部 session id）、末尾给行动建议；时间由系统通知呈现，
-  // 正文不重复时间戳。
-  const NOTIFY_KINDS: Record<string, { title: string; message: (detail: NotifyDetail) => string }> = {
-    ask: {
-      title: "DSH：等待审批",
-      message: (detail) => {
-        const lines = [];
-        if (detail.taskTitle) lines.push(`任务「${detail.taskTitle}」等待审批（工具「${prettyToolName(detail.tool)}」）`);
-        else lines.push(`工具「${prettyToolName(detail.tool)}」等待审批`);
-        if (detail.reason) lines.push(`理由：${detail.reason}`);
-        if (detail.remindMinutes) lines.push(`已等待 ${detail.remindMinutes} 分钟，仍在等待你的审批`);
-        lines.push("请到 DSH 界面确认或拒绝");
-        return lines.join("\n");
-      },
-    },
-    question: {
-      title: "DSH：向你提问",
-      message: (detail) => {
-        const lines = [];
-        if (detail.taskTitle) lines.push(`任务「${detail.taskTitle}」需要你回答`);
-        else lines.push("有提问需要你回答");
-        if (detail.question) lines.push(`问题：${detail.question}`);
-        lines.push("请到 DSH 界面回答");
-        return lines.join("\n");
-      },
-    },
-    done: {
-      title: "DSH：任务完成",
-      message: (detail) => {
-        if (detail.mergedCount) {
-          return `另有 ${detail.mergedCount} 个任务已完成（最近：${detail.taskTitle ?? "—"}）`;
-        }
-        const lines = [];
-        if (detail.taskTitle) lines.push(`任务「${detail.taskTitle}」已完成`);
-        else lines.push("后台任务已完成");
-        lines.push(`耗时：${formatDuration(detail.durationMs ?? 0)}`);
-        return lines.join("\n");
-      },
-    },
-    "subagent-done": {
-      title: "DSH：子任务完成",
-      message: (detail) => {
-        if (detail.mergedCount) {
-          return `另有 ${detail.mergedCount} 个子任务已完成（最近：${detail.taskTitle ?? "—"}）`;
-        }
-        const lines = [];
-        if (detail.taskTitle) lines.push(`子任务「${detail.taskTitle}」已完成`);
-        else lines.push("子任务已完成");
-        lines.push(`耗时：${formatDuration(detail.durationMs ?? 0)}`);
-        return lines.join("\n");
-      },
-    },
-    error: {
-      title: "DSH：任务出错",
-      message: (detail) => {
-        const lines = [];
-        if (detail.taskTitle) lines.push(`任务「${detail.taskTitle}」执行出错`);
-        else lines.push("任务执行出错");
-        if (detail.turn) lines.push(`第 ${detail.turn} 轮${detail.step ? `第 ${detail.step} 步` : ""}：${detail.message ?? ""}`);
-        else if (detail.message) lines.push(`错误：${detail.message}`);
-        if (detail.mergedCount) lines.push(`（窗口内另有 ${detail.mergedCount} 条同类错误）`);
-        if (detail.mergedErrors && detail.mergedErrors.length > 0) lines.push(`窗口内其他错误：${detail.mergedErrors.join(" / ")}`);
-        return lines.join("\n");
-      },
-    },
-    "turn-end": {
-      title: "DSH：轮次完成",
-      message: (detail) => {
-        const turn = detail.turn ? `第 ${detail.turn} 轮` : "";
-        if (detail.taskTitle) return `任务「${detail.taskTitle}」${turn}工作已完成`;
-        return `${turn}工作已完成`;
-      },
-    },
-  };
-
-  /**
-   * 通知历史落盘（jsonl 追加，滚动上限 HISTORY_LIMIT，tmp+rename 原子写）。
-   * fire-and-forget：不阻塞通知主流程；失败仅记 warn。
-   * 写队列串行化：多个事件并发触发时防止「读-改-写」竞态互相覆盖丢记录。
-   * 记录内容与通知文案同源（不含工具参数等敏感信息）。
-   */
-  let historyWriteChain: Promise<void> = Promise.resolve();
-  function appendHistory(entry: { ts: number; kind: string; title: string; message: string; suppressed?: string }) {
-    historyWriteChain = historyWriteChain.then(async () => {
-      try {
-        let lines: string[] = [];
-        try {
-          const text = await readFile(historyStore, "utf8");
-          lines = text.split("\n").filter(Boolean);
-        } catch {
-          // 文件不存在：从空列表开始
-        }
-        lines.push(JSON.stringify(entry));
-        // 按天自动清理（historyMaxAgeDays>0）：写入时剔除超过保留期的旧记录
-        if (current.historyMaxAgeDays > 0) {
-          const cutoff = entry.ts - current.historyMaxAgeDays * 86400000;
-          lines = lines.filter((line) => {
-            try {
-              const obj = JSON.parse(line);
-              return typeof obj.ts === "number" && obj.ts >= cutoff;
-            } catch {
-              return true; // 坏行保守保留
-            }
-          });
-        }
-        if (lines.length > HISTORY_LIMIT * 2) lines = lines.slice(-HISTORY_LIMIT);
-        const tmp = `${historyStore}.${process.pid}.${Date.now().toString(36)}.tmp`;
-        await writeFile(tmp, lines.join("\n") + "\n", "utf8");
-        await rename(tmp, historyStore);
-      } catch (error) {
-        ctx.logger.warn(`dsh-notifier: 历史记录写入失败: ${errorMessage(error)}`);
-      }
-    });
-  }
 
   /**
    * 通知入口：免打扰检查（被拦截也记录「未发出」历史）→ 系统通知 → SSE 广播 → 历史落盘。
@@ -781,8 +158,8 @@ export async function apply(ctx: Context, config: NotifierApplyConfig = {}): Pro
       return;
     }
     const payload = { type: "notify", kind, title, message, ts };
-    if (current.systemNotify) systemNotify(title, message);
-    if (current.browserNotify) broadcast(payload);
+    if (current.systemNotify) system.notify(title, message);
+    if (current.browserNotify) sse.broadcast(payload);
     ctx.logger.info(`dsh-notifier: ${kind} ${message.replace(/\n/g, " / ")}`);
     appendHistory({ ts, kind, title, message });
   }
@@ -1092,194 +469,26 @@ export async function apply(ctx: Context, config: NotifierApplyConfig = {}): Pro
     })
   );
 
-  // ------------------------------------------------------------ 路由
-
-  const configRoute: WebRoute = {
-    kind: "exact",
-    path: ROUTES.config,
-    handler: async (req, res) => {
-      if (!isLoopbackRequest(req)) return writeJson(res, 403, { error: "forbidden: loopback-only" });
-      if (req.method === "GET") {
-        writeJson(res, 200, current);
-        return;
-      }
-      if (req.method === "PUT") {
-        let body: unknown;
-        try {
-          body = await readBody(req, 16 * 1024);
-        } catch (error) {
-          const message = errorMessage(error);
-          if (message.includes("invalid JSON body")) {
-            // 非法 JSON：连接尚存，返回可读 400，避免请求「无响应挂起」
-            writeJson(res, 400, { error: `invalid JSON body: ${message}` });
-            return;
-          }
-          // 超限路径：readBody 已 reject 并 destroy 连接（防超大 body 占内存），
-          // socket 已断开无法再写响应——记录日志即可（shared/host-utils.js 语义）
-          ctx.logger.warn(`dsh-notifier: 配置请求体读取失败: ${message}`);
-          return;
-        }
-        current = normalizeConfig(body);
-        await saveConfig();
-        writeJson(res, 200, current);
-        return;
-      }
-      writeJson(res, 405, { error: `method not allowed: ${req.method}` });
-    },
-  };
-
-  const eventsRoute: WebRoute = {
-    kind: "exact",
-    path: ROUTES.events,
-    handler: (req, res) => {
-      if (!isLoopbackRequest(req)) {
-        writeJson(res, 403, { error: "forbidden: loopback-only" });
-        return;
-      }
-      if (req.method !== "GET") {
-        writeJson(res, 405, { error: `method not allowed: ${req.method}` });
-        return;
-      }
-      // ?since=<seq>：断线补拉——先回放滚动缓冲中 seq 更大的帧，再进入实时；
-      // 补拉独立于 /history（200 条截断），不丢尾部事件。
-      let since = 0;
-      try {
-        const raw = new URL(req.url ?? "/", "http://127.0.0.1").searchParams.get("since");
-        const parsed = raw === null ? 0 : Number(raw);
-        since = Number.isFinite(parsed) && parsed > 0 ? Math.floor(parsed) : 0;
-      } catch {
-        since = 0;
-      }
-      res.writeHead(200, {
-        "content-type": "text/event-stream",
-        "cache-control": "no-cache",
-        connection: "keep-alive",
-      });
-      res.write(": connected\n\n");
-      if (since > 0) {
-        for (const frame of recentFrames) {
-          if (frame.seq > since) {
-            try {
-              res.write(sseData(frame));
-            } catch {
-              break; // 连接已断
-            }
-          }
-        }
-      }
-      sseConnections.add(res);
-      res.on("close", () => {
-        sseConnections.delete(res);
-      });
-    },
-  };
-
-  /** 健康检查：插件是否加载、配置摘要、SSE 连接数。 */
-  const healthRoute: WebRoute = {
-    kind: "exact",
-    path: ROUTES.health,
-    handler: (req, res) => {
-      if (!isLoopbackRequest(req)) return writeJson(res, 403, { error: "forbidden: loopback-only" });
-      if (req.method !== "GET") return writeJson(res, 405, { error: `method not allowed: ${req.method}` });
-      writeJson(res, 200, {
-        ok: true,
-        plugin: "dsh-notifier",
-        config: {
-          notifyAsk: current.notifyAsk,
-          notifyQuestion: current.notifyQuestion,
-          notifyTaskDone: current.notifyTaskDone,
-          notifySubagentDone: current.notifySubagentDone,
-          notifyTaskError: current.notifyTaskError,
-          systemNotify: current.systemNotify,
-          browserNotify: current.browserNotify,
-          notifyWhenVisible: current.notifyWhenVisible,
-          notifySound: current.notifySound,
-          quietHours: current.quietHours,
-        },
-        sseConnections: sseConnections.size,
-      });
-    },
-  };
-
-  /**
-   * 测试通知：POST 触发一条测试通知（绕过免打扰，测试意图是验证通道本身）。
-   * 系统通知 + SSE 广播同时走一遍，客户端收到 kind=test 的帧无条件弹窗。
-   */
-  const testRoute: WebRoute = {
-    kind: "exact",
-    path: ROUTES.test,
-    handler: (req, res) => {
-      if (!isLoopbackRequest(req)) return writeJson(res, 403, { error: "forbidden: loopback-only" });
-      if (req.method !== "POST") return writeJson(res, 405, { error: `method not allowed: ${req.method}` });
-      const title = "DSH：测试通知";
-      const message = "通知链路工作正常（此通知来自测试按钮）";
-      const payload = { type: "notify", kind: "test", title, message, ts: Date.now() };
-      if (current.systemNotify) systemNotify(title, message);
-      if (current.browserNotify) broadcast(payload);
-      ctx.logger.info("dsh-notifier: test 测试通知已发送");
-      appendHistory({ ts: payload.ts, kind: "test", title, message });
-      writeJson(res, 200, { ok: true, sseConnections: sseConnections.size });
-    },
-  };
-
-  /** 通知历史：GET 返回最近记录（jsonl 尾部最多 HISTORY_LIMIT 条）。 */
-  const historyRoute: WebRoute = {
-    kind: "exact",
-    path: ROUTES.history,
-    handler: async (req, res) => {
-      if (!isLoopbackRequest(req)) return writeJson(res, 403, { error: "forbidden: loopback-only" });
-      if (req.method === "GET") {
-        const records: Array<Record<string, unknown>> = [];
-        try {
-          const text = await readFile(historyStore, "utf8");
-          const all = text.split("\n").filter(Boolean).slice(-HISTORY_LIMIT * 2);
-          const cutoff = current.historyMaxAgeDays > 0 ? Date.now() - current.historyMaxAgeDays * 86400000 : 0;
-          for (const line of all) {
-            try {
-              const obj = JSON.parse(line) as Record<string, unknown>;
-              if (cutoff > 0 && typeof obj.ts === "number" && obj.ts < cutoff) continue; // 超期过滤
-              records.push(obj);
-            } catch {
-              // 损坏行跳过
-            }
-          }
-          // 保守：只返回最近 HISTORY_LIMIT 条（写时已按行数与按天截断，此处兜底）
-          records.splice(0, Math.max(0, records.length - HISTORY_LIMIT));
-        } catch {
-          // 无历史文件：返回空列表
-        }
-        writeJson(res, 200, { ok: true, records });
-        return;
-      }
-      if (req.method === "DELETE") {
-        // 清空通知历史（写空文件原子化；返回被清空条数）
-        let removed = 0;
-        try {
-          const text = await readFile(historyStore, "utf8");
-          removed = text.split("\n").filter(Boolean).length;
-        } catch {
-          removed = 0;
-        }
-        try {
-          const tmp = `${historyStore}.${process.pid}.${Date.now().toString(36)}.clear.tmp`;
-          await writeFile(tmp, "", "utf8");
-          await rename(tmp, historyStore);
-        } catch (error) {
-          ctx.logger.warn(`dsh-notifier: 清空历史失败: ${errorMessage(error)}`);
-        }
-        writeJson(res, 200, { ok: true, removed });
-        return;
-      }
-      writeJson(res, 405, { error: `method not allowed: ${req.method}` });
-    },
-  };
+  // ------------------------------------------------------------ 路由注册
 
   // 关键：路由注册必须在 ctx.effect 内执行——apply 主体直接调用 ctx 服务
   // （webServer.register）会在 cordis isolate 链就绪前访问服务，触发
   // "Cyclic __proto__ value"（loader isolate 处理），导致插件激活失败。
+  const routes = buildRoutes({
+    getConfig: () => current,
+    putConfig: async (raw) => {
+      current = normalizeConfig(raw);
+      await saveConfig();
+      return current;
+    },
+    logger: ctx.logger,
+    sse,
+    system,
+    history: historyStore,
+  });
   const disposeRoutes = ctx.effect(
     () => {
-      const routeDisposers = [configRoute, eventsRoute, healthRoute, testRoute, historyRoute].map((route) => ctx.webServer.register(route));
+      const routeDisposers = routes.map((route) => ctx.webServer.register(route));
       return () => {
         for (const dispose of routeDisposers) {
           try {
@@ -1302,7 +511,7 @@ export async function apply(ctx: Context, config: NotifierApplyConfig = {}): Pro
   // 卸载时调用。若把 disposeRoutes() 写在 fn 主体，等于注册完立刻注销。
   ctx.effect(
     () => () => {
-      clearInterval(heartbeatTimer);
+      sse.dispose();
       // M4 定时器清理：审批提醒、完成风暴窗口
       for (const t of askRemindTimers.values()) clearTimeout(t);
       askRemindTimers.clear();
